@@ -8,6 +8,7 @@ dependencies:
     sdk: flutter
   meta: ^1.2.4
   background_fetch: ^1.0.0-nullsafety.1
+  collection: ^1.15.0-nullsafety.4
 
 flutter:
   plugin:
@@ -19,6 +20,6 @@ flutter:
         pluginClass: FLTBackgroundGeolocationPlugin
 
 environment:
-  sdk: ">=2.1.0 <3.0.0"
+  sdk: '>=2.12.0 <3.0.0'
   flutter: ">=1.10.0"
 
