@@ -11,27 +11,27 @@ part of flt_background_geolocation;
 ///
 class DeviceInfo {
   /// Device model
-  String model;
+  String? model;
 
   /// Device manufacturer
-  String manufacturer;
+  String? manufacturer;
 
   /// Device version
-  String version;
+  String? version;
 
   /// Platform iOS | Android
-  String platform;
+  String? platform;
 
   /// Framework `Flutter|ReactNative|Cordova`
-  String framework;
+  String? framework;
 
-  static DeviceInfo _instance;
+  static DeviceInfo? _instance;
 
   /// Retreive an singleton instance of `DeviceInfo`.
   static Future<DeviceInfo> getInstance() async {
     Completer completer = new Completer<DeviceInfo>();
     if (_instance == null) {
-      Map map = await _methodChannel.invokeMethod('getDeviceInfo');
+      Map map = await (_methodChannel.invokeMethod('getDeviceInfo') as FutureOr<Map<dynamic, dynamic>>);
       _instance = new DeviceInfo(
           model: map['model'],
           manufacturer: map['manufacturer'],
@@ -40,7 +40,7 @@ class DeviceInfo {
           framework: map['framework']);
     }
     completer.complete(_instance);
-    return completer.future;
+    return completer.future as FutureOr<DeviceInfo>;
   }
 
   DeviceInfo(
