@@ -6,22 +6,22 @@ part of flt_background_geolocation;
 ///
 class Sensors {
   /// Device platform `ios | android`.
-  String platform;
+  String? platform;
 
   /// `true` if this device has an accelerometer.
-  bool accelerometer;
+  bool? accelerometer;
 
   /// `true` if this device has a gyroscope.
-  bool gyroscope;
+  bool? gyroscope;
 
   /// `true` if this device has a magnetometer (compass).
-  bool magentometer;
+  bool? magentometer;
 
   /// __`[iOS only]`__ `true` if this iOS device has an **M7** co-processor (iPhone 5S+).
-  bool motionHardware;
+  bool? motionHardware;
 
   /// __`[Android only]`__ `true` if this Android device has a "significant motion" sensor.
-  bool significantMotion;
+  bool? significantMotion;
 
   Sensors(this.platform, this.accelerometer, this.gyroscope, this.magentometer,
       this.motionHardware, this.significantMotion);
